#9 [feat] 상세 정보 뷰 : [GET]

+++ cds-server/src/main/java/org/sopt/cdsserver/category/repository/CategoryJpaRepository.java
@@ -3,5 +3,5 @@ package org.sopt.cdsserver.category.repository;
 import org.sopt.cdsserver.category.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryJpaRepository extends JpaRepository<Category, Long> {
+public interface CategoryJpaRepository extends JpaRepository<Category, Integer> {
 }
